Modulo_2/semana_5/requirements.txt
# Database
psycopg2-binary==2.9.9

# Web Framework
fastapi==0.105.0
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Environment & Configuration
python-dotenv==1.0.0

# Data Generation & Validation
faker==20.1.0
email-validator==2.1.0

# Type Hints
typing-extensions==4.8.0

# Development & Testing (optional)
pytest==7.4.3
black==23.11.0
flake8==6.1.0

# HTTP Client (for testing)
httpx==0.25.2